12 "App Completed" prior to big CSS-update

+++ app/Http/Controllers/ChirpController.php
@@ -3,10 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Models\Chirp;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 
 class ChirpController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
@@ -49,10 +51,13 @@ class ChirpController extends Controller
             'message.max' => 'Chirps must be 255 characters or less.',
         ]);
 
-        Chirp::create([
-            'message' => $validated['message'],
-            'user_id' => null,
-        ]);
+        // Chirp::create([
+        //     'message' => $validated['message'],
+        //     'user_id' => null,
+        // ]);
+
+        // this works, even tho there is a compiler error for user()
+        auth()->user()->chirps()->create($validated);
 
         return redirect('/')->with('success', 'Your chirp has been posted!');
     }
@@ -78,6 +83,8 @@ class ChirpController extends Controller
      */
     public function edit(Chirp $chirp)
     {
+        $this->authorize('update', $chirp);
+
         return view('chirps.edit', compact('chirp'));
     }
 
@@ -86,6 +93,8 @@ class ChirpController extends Controller
      */
     public function update(Request $request, Chirp $chirp)
     {
+        $this->authorize('update', $chirp);
+
         $validated = $request->validate([
             'message' => 'required|string|max:255|min:5',
         ], [
@@ -104,6 +113,7 @@ class ChirpController extends Controller
     public function destroy(Chirp $chirp)
     {
         //$this->authorize('delete', $chirp);
+        $this->authorize('update', $chirp);
 
         $chirp->delete();
